LayoutTests/imported/w3c/web-platform-tests/css/css-pseudo/marker-computed-content-expected.txt


PASS Computed 'content' for list-item ::marker, variant default
PASS Computed 'content' for list-item ::marker, variant normal
PASS Computed 'content' for list-item ::marker, variant string
PASS Computed 'content' for list-item ::marker, variant image
FAIL Computed 'content' for list-item ::marker, variant none assert_equals: expected "none" but got "normal"
PASS Computed 'content' for non-list-item ::marker, variant default
PASS Computed 'content' for non-list-item ::marker, variant normal
PASS Computed 'content' for non-list-item ::marker, variant string
PASS Computed 'content' for non-list-item ::marker, variant image
FAIL Computed 'content' for non-list-item ::marker, variant none assert_equals: expected "none" but got "normal"
item
item
item
item
item
item
item
item
item
item
